requirements.txt
Django==1.10.4
django-assets==0.12
django-filter==1.0.1
django-taggit==0.21.3
djangorestframework==3.5.3
Markdown==2.6.7
Pillow==3.4.2
webassets==0.12.0
